+++ src/lib/holocene/collapsible-divider/collapsible-divider.stories.svelte
@@ -1,11 +1,11 @@
 <script lang="ts" context="module">
   import type { Meta } from '@storybook/svelte';
 
-  import Collapsible from './collapsible.svelte';
+  import CollapsibleDivider from './collapsible-divider.svelte';
 
   export const meta = {
-    title: 'Collapsible',
-    component: Collapsible,
+    title: 'Collapsible Divider',
+    component: CollapsibleDivider,
     args: {
       label: 'Expand me',
     },
@@ -13,7 +13,7 @@
     argTypes: {
       label: { name: 'Label', control: 'text' },
     },
-  } satisfies Meta<Collapsible>;
+  } satisfies Meta<CollapsibleDivider>;
 </script>
 
 <script lang="ts">
@@ -21,9 +21,9 @@
 </script>
 
 <Template let:args>
-  <Collapsible label={args.label}>
+  <CollapsibleDivider label={args.label}>
     <p>This is the content</p>
-  </Collapsible>
+  </CollapsibleDivider>
 </Template>
 
 <Story name="Default" />
